refactor(#486): split wifi.*/display.* into shared config providers
Epic #300. Extract the wifi.* and display.* config surface out of the
observer (wifi_observer/ObserverCli.cpp) into their own role-agnostic
providers so any WiFi/display-capable role uses one set of config keys, not
a reinvented-per-role set. Observer keeps mqtt.* only.

New src/helpers/config/:
  - WifiConfigProvider.{h,cpp}  -- prefix "wifi."; handlers relocated verbatim.
    Keeps its WifiBootstrap dependency for wifi.status (observer bring-up);
    what wifi.status means off-observer is #365's call (deliberate, deferred).
  - DisplayConfigProvider.{h,cpp} -- prefix "display."; handlers + appliers,
    PLUS the 4 display NVS accessors moved out of ConfigSchema (role-neutral
    offband_ui Preferences wrappers) so any display role links them without
    the observer schema. This is the owner-approved "decouple display now" half.

Each provider self-registers with the #364 dispatcher and declares its own
#366 key-prefix manifest; the observer's manifest drops wifi./display., so the
registration-time overlap detector guards the split.

ObserverCli: observerConfigSet/Get now handle mqtt.* only; dispatchObserverCli
(the observer _sys string CLI) still calls the moved handlers via the new
headers. ObserverCli.h forwards only DisplayConfigProvider.h so legacy includers
(companion main.cpp) resolve the display appliers + NVS accessors unchanged.

NO behavior change. Wire contract byte-compatible -- 141 client-visible reply
strings identical before vs after (device-side crashlog diagnostics excluded,
proven never written to a client reply). Builds: 4 observer envs + RAK_4631
(nRF52) = 5/5 SUCCESS; providers correctly NOT compiled into the nRF52 build.

Does NOT define the display cap bit / advertise / ver-code -- that is #365's
caps-byte surface (seam confirmed with #365 owner). CLI-verb->config-dispatch
bridge for non-observer string CLIs is #462.

Gemini adversarial review run; findings adjudicated in the PR body (MAJOR-1
pre-existing Offband bug -> #457; MAJOR-2 CrashLog relocation -> #350; MAJOR-3
string-CLI grammar -> #462; MINOR-3 fixed).
